Added the Tables slass, changed Basecommands to use them.
The Tables class dynamicly parses the 'tables.yaml' file
containing all the parameter tables used with the IMPBUS-2
protocoll (according to the IMPBUS-2 Developers manual). It
builds a class instance for every table inside the 'tables.yaml'
file. So to add or modify a IMPBUS-2 command all you have to do
is edit the YAML file. Well done boy!